backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asarPy==1.0.1
bcrypt==5.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
ecdsa==0.19.1
fastapi==0.119.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
